refactor: extract collect_stats/collect_doctor helpers
Move the per-project stats aggregation and doctor-checks logic into a
new stdout-clean module mindkeep._diagnostics. The CLI commands
_cmd_stats / _cmd_doctor become thin formatters over the helpers; the
forthcoming MCP read tools (#35) call the helpers directly so they can
return structured data without violating the stdio invariant
(DESIGN-v0.4.0 §3.4).

Notes:
- collect_stats deliberately omits the session_budget block (DESIGN
  §11: that's a CLI rendering concern). _cmd_stats appends it itself
  to keep --json output byte-identical for existing callers.
- collect_doctor accepts verbose=True/False to toggle per-check
  details (DESIGN §13). The CLI passes verbose=True; MCP defaults to
  False.
- Schema-version expectation can be overridden so the doctor test
  that simulates a newer-than-binary DB keeps working without us
  re-stamping meta.

No CLI behaviour change: existing tests for stats/doctor pass
unchanged.
